requirements.txt
# Core ML libraries
torch>=2.1.0,<2.3.0
torchaudio>=2.1.0,<2.3.0
torchvision>=0.16.0,<0.18.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Efficient fine-tuning
peft>=0.6.0
bitsandbytes>=0.41.0

# Faster inference
faster-whisper>=0.10.0
ctranslate2>=3.20.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
audiomentations>=1.4.0

# Lightning AI
lightning>=2.1.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Evaluation metrics
jiwer>=3.0.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
wandb>=0.16.0  # Optional: for experiment tracking

# Amharic text processing
python-ethiopic>=0.7.0  # For Ge'ez script handling
